refactor(calendar): split header into navigation and filtering components
Extracted filter and navigation logic from MonthCalendarHeader into new MonthCalendarFilters and MonthCalendarNavigation components for improved modularity and maintainability. Updated MonthCalendarHeader to use these new components, reducing its complexity and improving code organization.
